+++ fractal_tasks_core/dev/task_list.py
@@ -31,10 +31,10 @@ TASK_LIST = [
         meta={"cpus_per_task": 1, "mem": 4000},
     ),
     CompoundTask(
-        name="Maximum Intensity Projection HCS Plate",
+        name="Project Image (HCS Plate)",
         input_types={"is_3D": True},
         executable_init="tasks/copy_ome_zarr_hcs_plate.py",
-        executable="tasks/maximum_intensity_projection.py",
+        executable="tasks/projection.py",
         output_types={"is_3D": False},
         meta_init={"cpus_per_task": 1, "mem": 4000},
         meta={"cpus_per_task": 1, "mem": 4000},
@@ -46,13 +46,6 @@ TASK_LIST = [
         output_types=dict(illumination_corrected=True),
         meta={"cpus_per_task": 1, "mem": 4000},
     ),
-    # CompoundTask(
-    #     name="Illumination Correction (channel parallelized)",
-    #     input_types=dict(illumination_correction=False),
-    #     executable_init="illumination_correction_init.py",
-    #     executable="illumination_correction_compute.py",
-    #     output_types=dict(illumination_correction=True),
-    # ),
     ParallelTask(
         name="Cellpose Segmentation",
         executable="tasks/cellpose_segmentation.py",