+++ tests/tasks/test_registration.py
@@ -56,8 +56,8 @@ from fractal_tasks_core.tasks.init_group_by_well_for_multiplexing import (
     init_group_by_well_for_multiplexing,
 )
 from fractal_tasks_core.tasks.io_models import MultiplexingAcquisition
-from fractal_tasks_core.tasks.maximum_intensity_projection import (
-    maximum_intensity_projection,
+from fractal_tasks_core.tasks.projection import (
+    projection,
 )
 from fractal_tasks_core.utils import _split_well_path_image_path
 
@@ -296,10 +296,9 @@ def test_multiplexing_registration(
     # # MIP
     image_list_updates = []
     for image in parallelization_list:
-        image_list_updates += maximum_intensity_projection(
+        image_list_updates += projection(
             zarr_url=image["zarr_url"],
             init_args=image["init_args"],
-            overwrite=True,
         )["image_list_updates"]
 
     zarr_urls_2D = [image["zarr_url"] for image in image_list_updates]