+++ tests/tasks/test_workflows_cellpose_segmentation.py
@@ -54,8 +54,8 @@ from fractal_tasks_core.tasks.cellvoyager_to_ome_zarr_init import (
 from fractal_tasks_core.tasks.copy_ome_zarr_hcs_plate import (
     copy_ome_zarr_hcs_plate,
 )
-from fractal_tasks_core.tasks.maximum_intensity_projection import (
-    maximum_intensity_projection,
+from fractal_tasks_core.tasks.projection import (
+    projection,
 )
 from fractal_tasks_core.zarr_utils import OverwriteNotAllowedError
 
@@ -484,10 +484,9 @@ def test_workflow_with_per_well_labeling_2D(
     # MIP
     image_list_updates = []
     for image in parallelization_list:
-        image_list_updates += maximum_intensity_projection(
+        image_list_updates += projection(
             zarr_url=image["zarr_url"],
             init_args=image["init_args"],
-            overwrite=True,
         )["image_list_updates"]
 
     channel = CellposeChannel1InputModel(